Don't associate extension rdoc with Sequel module
RDoc doesn't differentiate file-level RDoc from RDoc for the first
module, so if you only have one RDoc comment block at the top
of the file, and then a class or module after it, the RDoc comment
will be used for both the file itself and the module.

Add empty comment above Sequel module so that the file level
comment isn't used for the Sequel module.