Handle composite primary keys when working around validation issues for one_to_(one|many) associations in nested_attributes plugin (Fixes #870)
When a model uses a composite primary key, pk is never nil.  The
previous code resulted in the foreign key in the associated object
getting set to an array, which could make validation of the
associated object fail.  This uses the 0 value temporarily in the
composite primary key case, just as it is used in the simple
primary key case.