Allow delayed evaluation blocks to take dataset as an argument
Previously, delayed evaluations could not take the dataset as an
argument, which means you couldn't have a delayed evaluation
that based its result on the dataset currently being literalized.

For example, if you wanted to have a dataset that had a column
reference that was always qualified to the dataset source, there
wasn't an easy way to do that.  Now you can do:

  d = Sequel.delay{|ds| Sequel.qualify(ds.first_source, :column)}
  ds = DB[:table].where(d)
  ds.sql # SELECT * FROM table WHERE table.column
  ds.from(:t) # SELECT * FROM t WHERE t.column

Note that this breaks backwards compatibility with
Dataset#delayed_evaluation_sql_append, which now takes the
DelayedEvaluation object instead of the underlying callable.