Support more array types by default in the pg_array extension, such as xml[] and uuid[]
This isn't comprehensive support, but it supports the most common
simple array types that weren't supported.  The remaining unsupported
types probably don't make sense to support as the scalar types
aren't readily usable.  For example, the geometric array types
should really be supported in a geometric extension that
adds support for both the scalar and array types, similar to the
other pg_* types.