Work around bug in old versions of MySQL when schema dumping a table with multiple timestamp columns (Fixes #882)
On MySQL, old versions of MySQL complain if you create a table
with multiple timestamp columns unless NULL is explicitly specified
(usually not needed in SQL as it is the default).  On MySQL, when
schema dumping a timestamp column in :same_db mode, explicitly
set :null=>true.  This will be overridden later to be :null=>false
if the column is actually NOT NULL.

This is only done in :same_db mode, as other databases don't have
this issue.

Note that this problem doesn't affect MySQL 5.6, but does affect
MySQL 5.1 and MariaDB 10.0.13.