Rip out fdbsql automatic query retry logic
If the database raises an error, Sequel should raise an error. If
an error is retryable, it's something that should be handled
internally to the database.

This restores the jdbc adapter code to the previous code.

This seems to fix the many_through_many 3 join table case.  Due
to the removal of the jdbc prepared statement override, it does
trigger an additional spec failure on jdbc/fdbsql, related to
stale prepared statements.  However, as that shouldn't be an
issue in production environments (issuing DDL that affects a
concurrent prepared statement), it's safe to just skip that
spec.