Make Fdbsql::Connection a subclass of PG::Connection
Previously, Fdbsql::Connection was a subclass of Object, and it
contained the PG::Connection.  However, this is inconsistent with
other Sequel adapters, which use the underlying connection object
directly.  In Sequel, DB.synchronize is supposed to yield the
underlying connection object, to make it easy to drop down directly
to the driver layer.  When the connection object is wrapped in a
separate object, this adds an additional, unwanted abstraction
layer, since users no longer get direct access to the underlying
connection object.

While here, move the check_version method to the Database and
call it server_version, since that's the method name most other
adapters use for it.