Make tactical_eager_loading plugin work for limited associations
Previously, Dataset#post_load was not called in this code path,
so the tactical_eager_loading metadata was not setup.  Call
post_load explicitly to make sure things are setup.

To avoid eagerly loading associations specified by the :eager
association option in this case this case, move the cascade
handling and integrate it into the elsif and else blocks.
Use the new placeholder literalizer with_dataset method in the
else block to clean up the code.