Make AssociationReflection#reciprocal not raise error if associated class contains association with invalid associated class
Previously, if model A has an association with model B, and
any associations in model B reference an invalid class,
calling reciprocal on the association reflection between model
A and model B would raise a NameError.

While this is good in that it alerts you to an error, I don't
think it's reciprocal's place to do so.

I'm on the fence about whether this is a good idea in general,
but for backwards compatibility this should probably go in. The
changes made since 4.14.0 make this more likely to happen in user
code.