Make validates_unique use a qualified primary key if the dataset is joined (Fixes #895)
If the dataset is joined, an unqualified primary key reference
could be ambiguous, and would be ambiguous when
class_table_inheritance is used with auto_validations.  This
qualifies the reference in such cases.

Note that there are other unqualified references still in the
query, but those can be qualified by using a :where option,
and don't affect the class_table_inheritance case.