Speed up Model.primary_key_hash and .qualified_primary_key_hash
The greater speedup is for .qualified_primary_key hash, since that
no longer calls primary_key_hash and then modifies the output.
Instead, it just uses a similar implementation.

In both cases, move the Symbol check first as that is the common
case, and raise an exception via else instead of at the top
of the method.

Also, for composite primary keys, use zip instead of
each_with_index, which should be slightly faster.