Stop converting '' default values to nil default values on MySQL
This has been done every since the MySQL-specific schema parsing
was added in 2.6.0.  I'm sure it was done for a good reason,
most likely because '' was used as a default value even when
the default was NULL.  However, this doesn't seem to be the
case anymore.  As it is useful to differentiate between a
NULL default and an empty string default, it's best to remove
the check.