Fix handling of Sequel::SQL::Blob instances in bound variables in the postgres adapter (Fixes #917)
This bug dates back to when support for bound variables was first
added.  Basically, escaping bytea values not only escapes the bytea
but also string escapes, which doubles apostrophes.  Since the
resulting value is not used SQL but as a bound variable, PostgreSQL
only unescapes the bytea and does not do the string unescaping first,
leading to the issue.

The pg gem supports using binary encoding for bytea values, which
allows you to use the values without encoding them, so just use
that.  In addition to being more correct, this is almost certainly
faster for large bytea values.

Add an integration test for this issue, to see if any other adapters
are affected.