Add Model#get_column_value and #set_column_value to get/set column values (Fixes #929)
This makes it easy to deal with model columns names that conflict
with existing ruby/Sequel method names.  This has been a long
standing issue with Sequel, but I couldn't think of a good way
to support it that didn't involve slowing down general use.
Instead of having these methods call send, they are aliases to
send.  This still may be a slight performance hit on ruby
implementations that optimize literal send calls, but at least
it isn't an additional ruby method call per column value.

This updates the core and all plugins that ship with Sequel to use
these new methods instead of send directly when send is used with
column names.

The associations support had options to work around the conflict
issue, but those should no longer be needed if you override
get_column_value and set_column_value appropriately.  For that
reason, the column naming conflict association options will
probably be going away in Sequel 5.