Add column_conflicts plugin to automatically handle columns that conflict with method names
This makes it easy to have Sequel automatically handle column
names that conflict with method names.  Just load the plugin
into the model, and it will override get_column_value and
set_column_value appropariately to set the conflict values
directly in the values hash instead of calling the method.