Add Model#cancel_action for canceling actions in before hooks, instead of having the hooks return false
Returning false from the before hooks to cancel the action has
always been clunky, but that was the way it has been since
January 1, 2008.  Inspired by ActiveRecord's deprecation of
this feature, add a cancel_action method which cancels the action.
ActiveRecord uses throw(:abort) for similar behavior, but I think
an instance method makes more sense in Sequel, considering that
hooks are already implemented interally using raise HookFailed.

The advantage to this over returning false is that in addition
to being simpler, you can also easily provide a specific
exception method, as opposed to the general "the before_* hook
failed".