Don't have the static cache plugin attempt to validate objects
When using the static_cache plugin with the auto_validations
plugin on a table with unique validations, when the static_cache
plugin froze the instances, it would attempt to validate each
instance, causing uniqueness checking queries to be sent to the
database.

Since you can't generally save a static_cache object, validation
isn't that important, and it is unlikely anyone is going to
staticly cache invalid objects and also care that valid? will
perform validation for them.