Update documentation
At least with sqlite3, the initial script from the README, can't be executed more than once.
This is occasioned when it opens a database that already contains a table with that name.
Even though this is documented, it is more appropriate to read it here too, the first contact with Sequel.