Correctly apply association limit when eager loadnig with an eager block using default limit strategy on some databases
On some databases, the default limit strategy when using an eager
block was :correlated_subquery, since that is valid for eager_graph.
But since eager doesn't support that, it should be changed to
:ruby.

This modifies AssociationReflection#apply_ruby_eager_limit_strategy
so that it doesn't check the configured strategy, it now assumes
that the :ruby strategy should be used.  This is technically a
break in backwards compatibility, as the method is public, but it
is unlikely to be called externally.