Use a conditional variable instead of busy waiting in the threaded connection pools on ruby 1.9+
If a connection is not already available and the maximum number of
connections is already in use, Sequel has always busy waited for
a connection to become available.  I'm guessing the historical
reason for this is when Sequel was created, ruby didn't support
a way to wait for a connection to be available with a timeout.

Starting in ruby 1.9, ConditionVariable#wait supports a timeout,
and this switches the connection pools to use that instead of busy
waiting.  On ruby 1.8, busy waiting is still used.

There are a couple advantages to this.  One is that it uses less
user and system CPU time, which may help machines that are
compute bottlenecked.  The second is that it makes acquiring a
connection more fair and much less likely to timeout when under
heavy resource contention.  With the busy wait approach:

1) thread A wants connection, none available, sleeps
2) thread B checks in connection
3) thread C wants connection, gets one checked in by B
4) thread A wants connection, none available, sleeps
5) thread C checks in connection
6) thread B wants connection, gets one checked in by C
7) go to 1)

This eventually results in A timing out.

With the condition variable approach, this doesn't happen, since
when thread B checks in the connection, it signals the condition
variable, so thread A wakes up immediately and grabs the
connection just checked in.