Use Class.new instead of empty exception classes
According to Eric Wong, Class.new instead of creating an empty
class saves about 200 bytes per class by avoiding creation of
a class definition scope.

However, if you don't create a class definition scope, then
the class's name is not set immediately, it will be delayed until
runtime and cached on the first call to name.  So call name on all
of the classes to ensure the name is cached.