Add spec_core_model rake task
This runs the core and model specs together.  This was the previous
behavior of rake spec but is no longer the default.  By default, I
think it is better to keep the core and model specs separate, to
ensure that the core doesn't rely on the model layer.  However,
it can be useful to run them together to make sure things still
work when all specs are loaded together, and this new rake task
allows for that.