Make auto_validation plugin use :from=>:values option to setup validations on the underlying columns
Previously, the auto_validations plugin didn't work correctly if
the model was transforming the column values.  Now it uses the new
:from=>:values option in validation_helpers to setup validations
on the underlying column, so it will work correctly even if the
model overrides the column setters, such as when the serialization
plugin is used.

While here, reduce hash allocations inside validate.