Make jdbc subadapters work if they issue queries while the subadapter is being loaded (Fixes #1022)
Before, if the subadapter issued queries while it was being loaded,
the queries would fail because the type convertor maps were not
setup yet.  Setup the default type convertor maps early, so that
queries work.  Still allow for overriding setup_type_convertor_map
in subadapters, which will be called after the subadapter is loaded.