Remove eager loading of cti subclasses
The method name chosen (eager) conflicts with the method name for
eagerly loading associations. Eager loading wasn't one of the existing
features of the class_table_inheritance plugin, and the reason hybrid
table inheritance was merged was to allow for single table inheritance
use, not for eager loading, so it's safer to just remove the feature.

Because of the use of lazy attributes, eager loading is already done
on a per-column as needed basis.  Actual eager loading of subclass
columns may be helpful in some cases, but it should probably use a new
API if it is introduced, and that API might be more appropriately
added to lazy_attributes.

Minor fixes and modifications while here.