Make Sequel default to ignoring NULL values when using IN/NOT IN with an empty array
Previously, Sequel considered NULL values when using IN/NOT IN
with an empty array.  It's arguable whether the previous
behavior was more correct.  From an SQL perspective, it kind of
is, since NULL input should generally result in NULL output,
but there can certainly be an argument than IN with an empty
array should be always false and NOT IN with an empty array
should be always true (no matter the column value).

The big problem with the previous behavior was that databases
didn't optimize the queries well, resulting in full table
scans.  That's the main reason for this behavior change.

A empty_array_consider_nulls extension is being added for
users that desire the previous behavior.