Allow preparing already prepared statements when emulating offsets
When determining the SQL to use for a prepared statement on a
dataset that is emulating offsets, prepare is called on a clone
of that dataset.  Since this check was just added to alert users
to a possible error, and not because it actually causes problems,
just skip the error in that situation.