Make offset emulation without order but with explicit selection handle ambiguous column names
MSSQL <2012 doesn't support offsets, so support has to be emulated
using the ROW_NUMBER window function.  Unfortunately, ROW_NUMBER
on MSSQL requires an ORDER BY clause, even there is no reason it
should.

Previously, if there wasn't an existing order, Sequel would do a
query to get the columns, and then order by the columns.  However,
that does not work if there are ambiguous columns.  Change things
so it looks at an existing selection and uses the columns from
the selection, which should be qualified appropriately.