Use range function constructors instead of casts for known range types in pg_range (Fixes #1066)
This allows the use of CURRENT_TIMESTAMP or other expressions inside
ranges.  Previously, you could only use literal time values.  Note
that you can only use a constructor function if you know the range
type, so for PGRange instances with unknown type, they are still
literalized to a string without a cast (this uses the pg_unknown type
in PostgreSQL).  Also, constructor functions don't appear to handle
empty ranges, so continue to use a cast for empty ranges.

This change may not work for custom range types that lack
constructor functions.  If it turns out people are actually using
such ranges, we can change this to only use a constructor function
for known range types.