Add Dataset#single_record! and #single_value! which don't require cloning the receiver
These should be faster due to the lack of cloning.  Additionally,
this optimizes single_value and with_sql_single_value to use
each{|_, v| return v} instead of values.first, which should be
faster as it doesn't create an array, and additionally works
correctly for model datasets.

Note that this breaks backwards compatibility for code that is
relying on with_sql_single_value on model datasets to return an
array in [:column, value] format.