Add before_after_save plugin, for refreshing created objects and resetting modified flag before calling after_create/update/save hooks
I'd like to make this the default behavior, but I think it's too
likely to break backwards compatibility for existing applications.
So add a plugin for the desired behavior, and that plugin will
become the default behavior in Sequel 5.

If someone really desires the current behavior and is not just
accidently relying on it, I can add a plugin for the current
behavior to Sequel 5.