Use a synchronized Hash for transactions
This solves two problems:

1. Unsynchronized usage of transaction options
2. The use of a global lock for accessing transactions

== Unsynchronized Usage

The method Sequel::Database#_trans is used to access transaction options
of a connection object. While initial access to these options was
synchronized, any further usage and/or modifications were not. Take the
following code for example:

    options = Sequel.synchronize { @transactions[connection] }

    options[:foo] = 'bar'

While the initial access (the options assignment) is synchronized,
nothing is preventing concurrent modifications when setting the :foo
key.

In case of the actual Sequel code the method
Sequel::Database#transaction_finished? suffered from this flaw. In
particular the following code was problematic:

      1: if supports_savepoints?
      2:   stack = _trans(conn)[:savepoints]
      3:   stack.pop
      4:   stack.empty?
      5: else
      6:   true
      7: end

On line 2 the access to the transactions options Hash _is_ synchronized,
but requesting the :savepoints key is _not_. This could sometimes result
in `[:savepoints]` being called on a nil object due to the options being
unmodified somewhere else (unsynchronized). In some cases this would
also lead to segmentation faults when using Sequel in combination with
the pg Gem on MRI.

== Global Locking

The method Sequel.synchronize uses a single mutex. This means that two
different Sequel::Database instances are not able to access their
transactions in parallel (when using this method).

By using a synchronized Hash we can remove the need for a global lock,
in turn allowing different Database instances to access their
transactions in parallel.

== Testing

To test all this I used the following script:

    require 'sequel'
    require 'securerandom'

    db_host = 'localhost'
    db_name = 'test'
    db_user = 'root'

    db_options = {test: true, max_connections: 10, encoding: 'UTF-8'}

    if defined?(JRuby)
      DB = Sequel.connect(
        "jdbc:postgresql://#{db_host}/#{db_name}?user=#{db_user}",
        db_options
      )
    else
      DB = Sequel.connect("postgres://#{db_user}@#{db_host}/#{db_name}", db_options)
    end

    puts 'Preparing...'

    DB.create_table!(:reviews) do
      primary_key :id

      String :title, null: false
    end

    DB.create_table!(:review_comments) do
      String  :comment
      Integer :review_id
    end

    class Review < Sequel::Model
    end

    class ReviewComment < Sequel::Model
    end

    100.times do
      rid = DB[:reviews].insert(:title => SecureRandom.hex)

      DB[:review_comments].insert(:comment => SecureRandom.hex, :review_id => rid)
    end

    stop = false

    puts 'Starting threads...'

    %w{INT TERM}.each do |signal|
      trap(signal) do
        stop = true
      end
    end

    threads = 10.times.map do
      Thread.new do
        loop do
          break if stop

          review = Review.select(:id).order { random{} }.first

          review.update(:title => SecureRandom.hex)

          DB.transaction(savepoint: true) do
            ReviewComment.where(:review_id => review.id).delete
          end

          DB.transaction(savepoint: true) do
            ReviewComment.insert(:review_id => review.id, :comment => SecureRandom.hex)
          end
        end
      end
    end

    threads.each(&:join)

Running this script on MRI will usually work, but sometimes it will
throw a segmentation fault. In our production systems it also throws
NoMethodErrors quite often, originating from the method
Sequel::Database#transaction_finished?.

The particular problems are highly timing sensitive meaning you could
run the above script a million times without getting any problems.
Having said that, on my local computer the above script segfaults pretty
quickly (= after 2 minutes or so) prior to these commits.