Run Database#table_exists? inside a savepoint if currently in a transaction and the database supports savepoints (Fixes #1100)
Some databases should a PostgreSQL will roll the savepoint or
transaction back if there is an error in any query inside the
savepoint/transaction.  Because table_exists? works by issuing
a query and checking for an exception, this meant it didn't work
inside transactions on PostgreSQL.

Running the problematic query in its own savepoint fixes the
issue, because in the worst case, only that savepoint is
rolled back, not the entire query.