Minor tweaks to uuid plugin
Don't require rubysl-securerandom, rubysl is already required for
the rbx platform and should include that.

Make it easy to override how UUIDs are created by adding a
Model#create_uuid private method.  As SecureRandom.uuid isn't
defined on ruby 1.8.7, this makes it possible to use the plugin
on ruby 1.8.7 if you override that method.

Use instance variables instead of let in the specs, similar to
existing Sequel specs.