Support using PostgreSQL-specific types in bound variables in the jdbc/postgresql adapter
I had previously tried to find a way to do this, but was unable to.
However, doing some more research on the problem last night, I
found that you can use PGObject to handle PostgreSQL-specific types.
Usually, this requires you specify the type, but you can use
unknown and it will work with all types, similar to how the
postgres adapters works.