Convert types in association_pks setters before saving them, instead of just before running queries
This works better when delaying the saving of pks, since it
ensures that the cached values are typecast correctly, so future
calls to the the getters return the correct types.

To simplify things, drop the convert_cpk_array private method,
and just have the convert_pk_array private method handle both
cases.