Make eager_each plugin handle eager loading for Dataset#first and similar methods
The eager_each plugin kind of worked for Dataset#first before Sequel
4.27.0, but the behavior was unspecified, and it only worked correctly
in the eager case, not in the eager_graph case.  Due to refactoring in
4.27.0, it no longer eager loaded in either case.

This brings eager loading for #first support back to eager_each, but it
fixes the eager_graph issue by doing a 2nd query in the eager_graph case
to eagerly all load associated records related to the main record.