Add a setter method for one_through_one associations
When one_through_one associations were added, a setter method was
not added because few applications would need it, so I decided to
wait until someone asked for it.  We got a request to add a setter
method a couple months back, so this adds the necessary code.

The method is a little different than the other setter methods,
as it has to handle 5 separate cases:

* association = nil, no entry in join table (nothing)
* association = nil, entry in join table (DELETE)
* association = value, no entry in join table (INSERT)
* association = value, same entry in join table (nothing)
* association = value, different entry in join table (UPDATE)

While here I restructed the integration tests a bit to make them
more explicit.  I also changed module inclusion in the integration
tests to avoid the running of many redundant tests.

When adding the specs, I found that one_to_one had a duplicate
spec, which this removes.