Make association_pks setter handle empty array correctly when :delay_pks is set
If :delay_pks was set but the resulting array was empty, Sequel
was not making any modifications, making it impossible to use
the association_pks setter to remove all currently associated
objects.  This has been true since the :delay_pks option was
added.

Start handling this case by not skipping modification, and change
the setters to just call the remove_all method in such a case,
since that will perform better.

While here, remove automatic conversion of setter method argument
to array, as it could hide bugs.