Support :association_pks_nil association option in association_pks setter for determining how to handle nil
In earlier Sequel versions, if you provided nil, it would
result in an error, either immediately if :delay_pks was not set,
or when saving if :delay_pks was set.

In Sequel 4.31.0, if you provided nil and :delay_pks was not set,
it would probably remove all associated rows.  If :delay_pks was set,
it was ignored, as it was turned into an empty array and empty arrays
where ignored.

As of the last commit, behavior was changed back to the pre-4.31.0
behavior, always raising an error.  However, users may want one
of the two behaviors used by 4.31.0.  They may want to ignore
nil values, or they may want to remove all existing values.

One possible use case for removing existing values on nil is when
using association_pks with a rack application. If you have separate
checkbox inputs for each associated object and none are checked,
no params will be sent for the parameter name, so the rack params
value for the input name will be nil.