Add no_auto_literal_strings extension, which makes SQL injection vulnerabilities less likely
Inadvertent use of user input in plain strings passed to filter
methods is probably the #1 vulnerability in general Sequel usage.
By raising an error in the plain string case, it will alert users
when they may be doing something that makes them vulnerable.

I'm not sure this is a good default behavior, as there is some
harm to usability, but I recommend usage of this extension in
all security-sensitive applications.