Allow schema_dumper to handle :qualify=>true option on PostgreSQL
This removes some error checking code that disallowed the use of
SQL::QualifiedIdentifier instances as table names.

This does change the output slightly when using an SQL::Identifier,
since it is no longer automatically converted to a string, but
I think that is an acceptable change as the migration will still
function the same way.