Sequel::Model no longer swallows many errors when subclassing or setting datasets (Fixes #1160)
Historically Sequel has been pretty loose with swallowing errors
during model subclassing and dataset setting.  This was because
I wasn't sure when adding features years ago whether it was safe,
and I didn't want to risk breaking things.

I have a better idea now which things are safe, and while I still
think this does risk breaking things, in most cases it's going
to point out something already broken.

Specific changes:

* Model.inherited no longer rescues exceptions raised by set_dataset
* When subclassing a model that has a dataset, the columns and
  schema are just copied from the superclass
* set_dataset no longer wraps get_db_schema call in
  check_non_connection_error, as get_db_schema uses that internally
* Only Sequel::Error is rescued in check_non_connection_error

This also makes it so the columns call inside get_db_schema
is wrapped in a savepoint if inside a transaction and savepoints
are supported.  This fixes issues when creating models with
invalid tables inside a transaction on PostgreSQL.