The boolean_readers, boolean_subsets, and class_table_inheritance plugins no longer do blind rescues (Fixes #1162)
This changes check_non_connection_error to use a savepoint around
possible errors if inside a transaction, instead of just doing
that around the call to columns in get_db_schema.  This allows it
to handle database errors that abort transactions in db_schema.

The plugins now wrap calls to columns in check_non_connection_error,
instead of using columns rescue nil.

Add an integration test for subclassing a model with an invalid
table inside a transaction, making sure the transaction still works.