Add Database#transaction :savepoint=>:only option, for only creating a savepoint if already inside a transaction
The use case for this is when running queries that may cause a
database error, where you have code that rescues the error.

On at least PostgreSQL, a database error inside a transaction
breaks the rest of the transaction.  The exception in ruby will
be rescued, but the database connection state is broken. By
wrapping such calls in a savepoint, it automatically handles
database errors by only rolling back the savepoint.