Add support for :dataset_associations_join association option to dataset_associations plugin, for making resulting datasets have appropriate joins
Previously, associations that used JOINs and had a select, order,
or filter that depended on the joined table did not work
correctly with dataset associations.  By using the
:dataset_associations_join association option, the dataset
association will now set up similar joins in the resulting dataset,
allowing the select, order, or filter conditions to work as
expected.