Fix race conditions in the :preconnect=>:concurrently support
The original patch was not thread safe, as it called checkin_connection
inside the spawned threads, which could modify the available_connections
hash concurrently.  This caused non-deterministic spec failures, and
possibly worse.

Change the feature so that the connections are made concurrently in
separate threads, but checked in in serial in the main thread, avoiding
this issue.