Treat :Bignum as a generic type, to support 64-bit integers on ruby 2.4+, where Bignum == Integer
In ruby 2.4, Fixnum and Bignum are just aliases for Integer.
This means that code like:

  DB.add_column :foo, Bignum
  # and
  DB.get(Sequel.cast('1', Bignum))

will start to change meaning in ruby 2.4.  Previously, Bignum
here would reference a 64-bit integer type, but with the changes
in ruby 2.4, it will start to reference a 32-bit integer type.

Unfortunately, there is no good way to fix this in a backwards
compatible manner, without changing ::Bignum to be different
than ::Integer.  That's the breaks when relying on constants you
don't control.

I've considered a few different approaches to supporting a
generic 64-bit Integer type.  One was adding a Sequel::Bignum
class and using that instead, but that's kind of long.  Another
was adding a Bignum() method to some classes, which is less long
but unfortunately does not handle all cases.

This approach seemed the simplest, even though it creates an
inconsistency, in that the :Bignum symbol is treated as generic,
while all other symbols are treated as specific.  It also isn't
backwards compatible with anyone using :Bignum to create a specific
type, but I doubt anyone is doing that.  From a cost/benefit
perspective, I think this is the best fix.

Note that the changes in ruby 2.4 will affect old migrations, so
after upgrading to a version of Sequel that includes this commit,
you would need to modify all migrations to replace Bignum with
:Bignum in the cases where it is used as a generic type.