Make :rollback=>:always inside a transaction use a savepoint automatically if supported (Fixes #1193)
If savepoints are not supported, using :rollback=>:always inside
a transaction raises an Error.

I decided to dup the options hash at the top of #transaction. This
is a small performance hit in most cases, but considering the
overhead of #transaction, I doubt it is measurable on a real
database.

Refactor some of the related code while here.