add YARD configuration
While this project natively uses RDoc documentation format, it is possible to
use it with [YARD][1] without any major modifications.

One of the reasons why it might be useful is usage with offline documentation
tools that use YARD to generate HTML documentation from both RDoc and YARD input.
One popular example of such applications is [Dash][2].

Prior to this commit, the documentation converted that way included only
information that was embedded directly in the code, and all useful documents
from `doc/*.rdoc` were missing.

[1]: http://yardoc.org/
[2]: https://kapeli.com/dash