Support operator validation in constraint_validations for <, <=, >, and >= operators with string and integer arguments
When the operator validation was added to validation helpers, there
wasn't a matching validation added to constraint validations.

Unfortunately, the constraint validations support is limited to
only the 4 basic inequality validations and only string or integer
arguments.  It's possible that additional operators and argument
types can be supported later if needed.