Make connection pool remove connections for disconnect errors not raised as DatabaseDisconnectError
This makes the connection pool also recognize error classes that
the adapter knows represent database errors, and call
disconnect_error? on the Database object to determine if the
exception is recognized as a disconnect error.  If the exception
is recognized a disconnect error, the connection is removed from
the pool, just as if DatabaseDisconnectError was raised.

Note that this does no conversion of error classes, the original
exception is always still raised.  However, it should ensure that
as long as an exception is recognized as a disconnect error by
the adapter, the connection will be removed from the pool.  This
fixes cases where the adapter was not checking for disconnect
errors and raising them as DatabaseDisconnectErrors, and the
bad connections remained in the pool until the application was
restarted.