Support ** operator for exponentiation on expressions, similar to +, -, *, and /
Exponentiation is not used nearly as much as the other mathmatical
operators, but since ruby support it via **, it would make sense
for Sequel to support it as well.

One additional benefit is that unlike +, -, *, and /, exponentation
is not completely standardized.  Most databases support the power
function, but access uses the ^ operator, derby requires a
combination of exp and ln (with some floating point error added),
and SQLite doesn't support exponentiation at all, though Sequel
can emulate it via multiplication if the exponent is an integer.